changes id espec and implements first functional client connection, by adding framing to the serial

+++ src/espMeshFlood/protocol/mesh_protocol.h
@@ -96,7 +96,7 @@ private:
      * @brief Generate unique message ID
      * @return New message ID
      */
-    uint64_t generate_message_id();
+    uint32_t generate_message_id();
 
     /**
      * @brief Schedule retransmission of a message
